// packages/datadog-api-client-v1/models/SyntheticsBrowserError.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SyntheticsBrowserErrorType } from './SyntheticsBrowserErrorType';
/**
* Error response object for a browser test.
*/
export declare class SyntheticsBrowserError {
    /**
    * Description of the error.
    */
    'description': string;
    /**
    * Name of the error.
    */
    'name': string;
    /**
    * Status Code of the error.
    */
    'statusCode'?: number;
    'type': SyntheticsBrowserErrorType;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsBrowserError;
    static serialize(data: SyntheticsBrowserError): {
        [key: string]: any;
    };
    constructor();
}